refactor(main): extract attachment dialog handlers
